+++ src/main/java/pl/dahdev/bloomfilters/TripleHashBloomFilter.java
@@ -162,18 +162,18 @@ public class TripleHashBloomFilter<E> extends AbstractBloomFilter<E> {
      */
     @Override
     protected int[] createHashes(byte[] bytes, int numberOfHash) {
-        long valueA, valueB, valueC;
+        int valueA, valueB, valueC;
         int[] hashes = new int[numberOfHash];
 
         valueA = getValueFromGeneratedHash(bytes, firstHash);
         valueB = getValueFromGeneratedHash(bytes, secondHash);
         valueC = getValueFromGeneratedHash(bytes, thirdHash);
 
-        hashes[0] = (int) valueA;
+        hashes[0] = valueA;
         for (int i = 0; i < numberOfHash; i++) {
             valueA = (valueA + valueB) % size;
             valueB = (valueB + valueC) % size;
-            hashes[i] = (int) valueA;
+            hashes[i] = valueA;
         }
         return hashes;
     }
